backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
httpx
google-genai
